requirements-dev.txt
ruff==0.8.2
pytest-servers[s3]
pytest-cov
coverage[toml]
